Added the power BI and SQL files for the two projects.
Power BI file has the cleaned, transformed data already loaded into power BI and analysis already done on the NREGA csv file, the various pages in the pbix file show the visualizations made on the fields in the NREGA dataset. They provide valuable insights on the effectiveness of the program, and key recommendations to improve it.
The sql file has a few queries that were run on the Corona dataset, the dataset has columns like confirmed, recovered and death columns, these were very sensitive columns that provided insights on the effectiveness of the disease, queries were run on the dataset and the various results are displayed in the presentation.